IGPS 4.2.9                 - Jan 16th      2025 	============ 	- TIP_FW 0.7.8 L0 0.6.7 L1 		https://github.com/Nuvoton-Israel/npcm8xx-tip-fw/releases/tag/TIP_FW_0.7.8_L0_0.6.7_L1 		* Add command for DRBG from BMC to TIP. 		* Add to TIP_FW 'SystemControlFlags' flag to enable\disable features in TIP_FW: currently 3 bits: tip_recovery_force, tip_start_wd, tip_support_encryption 		* Disable flash protection. 		* Disable WOL.
	- Add to TIP_FW L1 XML 'SystemControlFlags' flag to enable\disable features in TIP_FW
	- Fixed default OTP LMS key0.
	- Optimize signing time for LMS.
	- Combo 1 : If LMS enabled all components are signed.
	- Add support for Meta.

+++ py_scripts/ImageGeneration/IGPS_files.py
@@ -364,15 +364,19 @@ bin_registers_uboot     = os.path.join("output_binaries", "regs", "registers_ubo
 
 from  .key_setting_edit_me import *
 
-def CheckIfFileExistsAndMove(srcFile, toFolder):
+def CheckIfFileExistsAndMove(srcFile, toFolder, ForceCopy = True):
 	dstFile = os.path.join(toFolder, os.path.split(srcFile)[-1])
 	
 	if (os.path.isfile(srcFile) == False):
 		print(("\033[91m" + "CheckIfFileExistsAndMove   Error: " +  srcFile + " file is missing\n\n" + "\033[97m"))
 		return
 	
-	print(("     Copy " +  srcFile + " to " + toFolder))
-	
 	if os.path.isfile(dstFile):
-		os.remove(dstFile)
+		if (ForceCopy == False):
+			print("     Destination file " + dstFile + " already exists. Copy operation skipped.")
+			return
+		else:
+			os.remove(dstFile)
+	
+	print("     Copy " + srcFile + " to " + toFolder)
 	copy(srcFile, toFolder)
